refactor: reorganize component structure and update imports; remove unused components
