+++ src/app/(home)/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 
-import Footer from "@/components/footer"
-import Header from "@/components/header";
+import Footer from "@/components/layout/footer"
+import Header from "@/components/layout/header";
 
 import "@/app/globals.css";
