+++ src/components/layout/header.tsx
@@ -5,7 +5,7 @@ import Link from "next/link"
 import React from "react"
 import Image from "next/image"
 import { ModeToggle } from "@/components/theme/mode-toggle"
-import { VisitGitHub } from "@/components/visit-github"
+import { VisitGitHub } from "@/components/layout/github"
 import { Button } from "@/components/ui/button"
 import { Menu } from "lucide-react"
 import {
