+++ src/lib/menu-list.ts
@@ -27,6 +27,7 @@ type Group = {
   menus: Menu[];
 };
 
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 export function getMenuList(pathname: string): Group[] {
   return [
     {
